+++ src/__tests__/ghost.test.js
@@ -1,7 +1,7 @@
 jest.mock('3id-resolver', () => {
   const { didResolverMock } = require('../__mocks__/3ID')
   return {
-    getResolver: () => ({'3': didResolverMock})
+    getResolver: () => ({ '3': didResolverMock })
   }
 })
 
@@ -13,27 +13,34 @@ const { threeIDMockFactory } = require('../__mocks__/3ID')
 const DID1 = 'did:3:zdpuAsaK9YsqpphSBeQvfrKAjs8kF7vUX4Y3kMkMRgEQigzCt'
 const DID2 = 'did:3:zdpuB2DcKQKNBDz3difEYxjTupsho5VuPCLgRbRunXqhmrJaX'
 const DID3 = 'did:3:zdpuAvKY6Noex9nRMp1pR8dHauT5Rn4yhVuLq9bAdz4pp5oRd'
+const DID4 = 'did:3:bafyreifgwegkzk6rxemfck2vfm2zpw223hkl37eqvtms6gdwu2ejnbkrly'
 
 const CHAT_NAME = '3box.chat.somespace.name'
 
-const THREEID1_MOCK = threeIDMockFactory(DID1);
-const THREEID2_MOCK = threeIDMockFactory(DID2);
-const THREEID3_MOCK = threeIDMockFactory(DID3);
+const THREEID1_MOCK = threeIDMockFactory(DID1)
+const THREEID2_MOCK = threeIDMockFactory(DID2)
+const THREEID3_MOCK = threeIDMockFactory(DID3)
+const THREEID4_MOCK = threeIDMockFactory(DID4)
 
 describe('Ghost Chat', () => {
   let ipfs
   let chat
   let chat2
   let peer
-  let user = 0
+  const user = 0
   // jest.setTimeout(2000000)
 
   beforeAll(async () => {
-    ipfs = await utils.initIPFS(11);
+    ipfs = await utils.initIPFS(11)
   })
 
   beforeEach(async () => {
     jest.setTimeout(30000)
+  })
+
+  afterAll(async () => {
+    await chat.close()
+    await ipfs.stop()
   })
 
   it('creates chat correctly', async () => {
@@ -46,16 +53,19 @@ describe('Ghost Chat', () => {
     expect(chat._3id).toEqual(THREEID1_MOCK)
   })
 
-  it('should catch messages', async (done) => {
-    chat.onUpdate(async ({ type, author, message }) => {
-      if (type == 'chat') {
-        expect(author).toEqual(DID1)
-        expect(chat.getPosts()).not.toEqual([])
-        expect(chat.getPosts()).toBeDefined()
-        done()
-      }
+  it('should catch messages', async () => {
+    const postPromise = new Promise(resolve => {
+      chat.onUpdate(async ({ type, author, message }) => {
+        if (type == 'chat') {
+          expect(author).toEqual(DID1)
+          expect(chat.getPosts()).not.toEqual([])
+          expect(chat.getPosts()).toBeDefined()
+          resolve()
+        }
+      })
     })
     await chat.post('hello')
+    await postPromise
   })
 
   describe('multi user interaction', () => {
@@ -64,6 +74,13 @@ describe('Ghost Chat', () => {
 
     beforeAll(async () => {
       ipfs2 = await utils.initIPFS(12)
+      let ipfsMultiAddr = (await ipfs.id()).addresses[0]
+      await ipfs2.swarm.connect(ipfsMultiAddr)
+    })
+
+    afterAll(async () => {
+      await chat2.close()
+      await ipfs2.stop()
     })
 
     it('creates second chat correctly', async () => {
@@ -76,7 +93,7 @@ describe('Ghost Chat', () => {
           resolve()
         })
       })
-      chat2 = new GhostThread(CHAT_NAME, { ipfs: ipfs2 });
+      chat2 = new GhostThread(CHAT_NAME, { ipfs: ipfs2 })
       const c2Promise = new Promise(resolve => {
         chat2.on('user-joined', async (_event, did, peerId) => {
           expect(_event).toEqual('joined')
@@ -153,26 +170,32 @@ describe('Ghost Chat', () => {
       })
       await chat._requestBacklog()
     })
-
-    afterAll(async () => {
-      await chat2.close()
-      await utils.stopIPFS(ipfs2, 12)
-    })
   })
 
   describe('ghost filter tests', () => {
     let chat3
     let ipfs3
-    let filter = (payload, issuer, from) => {
+    const filter = (payload, issuer, from) => {
       if (issuer == DID1) {
-          return false
+        return false
       }
       return true
     }
 
+    beforeAll(async () => {
+      ipfs3 = await utils.initIPFS(12)
+      let ipfsMultiAddr = (await ipfs.id()).addresses[0]
+      await ipfs3.swarm.connect(ipfsMultiAddr)
+    })
+
+    afterAll(async () => {
+      await chat3.close()
+      await ipfs3.stop()
+    })
+
     it('creates third chat correctly', async (done) => {
       chat.removeAllListeners()
-      chat3 = new GhostThread(CHAT_NAME, { ipfs: ipfs3 }, { ghostFilters: [filter] });
+      chat3 = new GhostThread(CHAT_NAME, { ipfs: ipfs3 }, { ghostFilters: [filter] })
       chat3._set3id(THREEID3_MOCK)
       expect(chat3._name).toEqual(CHAT_NAME)
       expect(chat3._3id).toEqual(THREEID3_MOCK)
@@ -201,19 +224,82 @@ describe('Ghost Chat', () => {
       })
       await chat.post('wide')
     })
+  })
+
+  describe('Ghost Pinbot interaction', () => {
+    let ipfs4
+    let chat4
+    let ipfsGhost
+    let chatGhost
+    let ghostMultiaddr
+
+    const GHOST_PINBOT_CHAT_NAME = '3box.chat.somespace-ghost.name-ghost'
 
     beforeAll(async () => {
-      ipfs3 = await utils.initIPFS(12)
+      ipfs4 = await utils.initIPFS(14)
+      ipfsGhost = await utils.initIPFS(15)
+
+      let ghostMultiaddr = (await ipfsGhost.id()).addresses[0]
+      await ipfs4.swarm.connect(ghostMultiaddr)
     })
 
     afterAll(async () => {
-      await chat3.close()
-      return utils.stopIPFS(ipfs3, 12)
+      await chat4.close()
+      await chatGhost.close()
+
+      await ipfs4.stop()
+      return ipfsGhost.stop()
+    })
+
+    it('creates Ghost Pinbot correctly', async () => {
+      chatGhost = new GhostThread(GHOST_PINBOT_CHAT_NAME, { ipfs: ipfsGhost })
+      expect(chatGhost._name).toEqual(GHOST_PINBOT_CHAT_NAME)
+      expect(chatGhost.listMembers).toBeDefined()
+      expect(chatGhost.getPosts()).toBeDefined()
+      expect(chatGhost.isGhost).toBeTruthy()
+    })
+
+    it('creates fourth chat correctly', async (done) => {
+      // checks if chat4 joined properly
+      const gPromise = new Promise(resolve => {
+        chatGhost.onUpdate(async (msg) => {
+          if (msg.type === 'backlog') {
+            resolve()
+          }
+        })
+
+        chatGhost.on('backlog_response', async (_event, did, peerId) => {
+          expect(_event).toEqual('joined')
+          const members = await chatGhost.listMembers()
+          expect(members).toEqual(expect.arrayContaining([DID4]))
+          resolve()
+        })
+      })
+
+      chat4 = new GhostThread(GHOST_PINBOT_CHAT_NAME, { ipfs: ipfs4 }, {
+        ghostPinbot: ghostMultiaddr
+      })
+
+      chat4.onUpdate(async ({ type, author, message }) => {
+        if (type === 'backlog') {
+          expect(author).toBeUndefined()
+          const members4 = await chat4.listMembers()
+          expect(members4).toEqual([])
+          done()
+        }
+      })
+
+      chat4._set3id(THREEID4_MOCK)
+      expect(chat4._name).toEqual(GHOST_PINBOT_CHAT_NAME)
+      expect(chat4._3id).toEqual(THREEID4_MOCK)
+      expect(chat4.listMembers()).toBeDefined()
+      expect(chat4.getPosts()).toBeDefined()
+      await gPromise
     })
   })
 
   afterAll(async () => {
-    await chat1.close()
+    await chat.close()
     return utils.stopIPFS(ipfs, 11)
   })
 })